fligner {matrixTests}	R Documentation

Fligner-Killeen test

Description

Performs the Fligner-Killeen test of homogeneity of variances (with median centering of the groups) on each row/column of the input matrix.

Usage

row_flignerkilleen(x, g)

col_flignerkilleen(x, g)

Arguments

x
numeric matrix.

g
a vector specifying group membership for each observation of x.

Details

NA values are always ommited. If values are missing for a whole group - that group is discarded. Groups with only one observation are also discarded.

row_flignerkilleen(x, g) - Fligner-Killeen test on rows. col_flignerkilleen(x, g) - Fligner-Killeen test on columns.

Results should be the same as as running fligner.test(x, g) on every row (or column) of x.

Value

a data.frame where each row contains the results of the Fligner-Killeen test performed on the corresponding row/column of x.

Each row contains the following information (in order):
1. obs.tot - total number of observations
2. obs.groups - number of groups
3. df - degrees of freedom
4. statistic - squared statistic
5. pvalue - p-value

Author(s)

Karolis Koncevičius

See Also

fligner.test()

Examples

col_flignerkilleen(iris[,1:4], iris$Species)
row_flignerkilleen(t(iris[,1:4]), iris$Species)

[Package matrixTests version 0.2.3.1 Index]