waerden {matrixTests}	R Documentation

Van der Waerden test

Description

Performs van der Waerden test on each row/column of the input matrix.

Usage

row_waerden(x, g)

col_waerden(x, g)

Arguments

x
numeric matrix.

g
a vector specifying group membership for each observation of x.

Details

row_waerden(x, g) - van der Waerden test on rows. col_waerden(x, g) - van det Waerden test on columns.

Value

a data.frame where each row contains the results of van det Waerden test performed on the corresponding row/column of x.

Each row contains the following information (in order):
1. obs.tot - total number of observations
2. obs.groups - number of groups
3. df - degrees of freedome
4. statistic - van det Waerden chi-squared statistic
5. pvalue - p.value

Author(s)

Karolis Koncevičius

See Also

vanWaerdenTest, row_oneway_equalvar, row_kruskalwallis

Examples

col_waerden(iris[,1:4], iris$Species)
row_waerden(t(iris[,1:4]), iris$Species)

[Package matrixTests version 0.2.3.1 Index]